Lay groundwork for EAV utils mini-module: raw datatypes and converters
EAV stands for the Entity–Attribute–Value model, which is what we're
essentially applying by supporting sparse potentially heterogenous
attributes.

Nothing yet utilizes these datatypes/converters. Django-specific
utilization will be in forthcoming commits.

Note:
This commit assumes that the underlying database field stores a string,
and that Django *mostly* handles the type conversion when asked to store
a non-string into that field.

The purpose of converters in this code is to reliably convert...
- from string OR nonstring value
- to nonstring / usable Python value

In the latter case (nonstring -> nonstring), the intent is to be
idempotent to have a normalized value via the converter.

A Django-centric way to accomplish this might be subclassing Django's
built-in Field classes, but this groundwork is intended to be more
basic - allowing flexibility in how the database storage is
accomplished.