admin: improve datatype field display, for customized boolean type
A workaround was necessary, as Django 2.2 would not seem to allow
get_FOO_display() to be overrided. It likely is a method added by the
model metaclass, and thus not something that can be easily customized.

By contrast, Django 3.0 docs add a note that it is typically
overrideable, so likely there was an improvement between 2.2 and 3.0.

In lieu of that, a datatype_display property has been added to the
model, and an AdminMixin created to use that property automatically
instead of datatype. This provides at least *something* that can be
overrideable, albeit requiring a customization to the Admin class.