admin: support value entry using custom boolean choices
Depending on the user, it could be unnatural if the admin input only
accepted true/false rather than the customized wording. Thus we support
it here.

Outside of the admin entry, we need not invoke this shim. The customized
choice can be for display only, and input would be mediated by
form/template so that raw True/False would be handled behind the scenes
(or via some other strategy).